pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.120.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.0",
    "langchain-mongodb>=0.7.1",
    "lxml-html-clean>=0.4.3",
    "newspaper3k>=0.2.8",
    "pymongo>=4.15.3",
    "pypdf>=6.1.2",
]
